+++ src/main/java/frc/robot/RobotContainer.java
@@ -10,6 +10,8 @@ import frc.robot.commands.ExampleCommand;
 import frc.robot.commands.StartIntake;
 import frc.robot.subsystems.ExampleSubsystem;
 import frc.robot.subsystems.Intake;
+import frc.robot.subsystems.IntakePivot;
+import frc.robot.subsystems.IntakeRollers;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
@@ -25,7 +27,10 @@ import edu.wpi.first.wpilibj2.command.button.Trigger;
 public class RobotContainer {
   // The robot's subsystems and commands are defined here...
   private final ExampleSubsystem m_exampleSubsystem = new ExampleSubsystem();
-  private final Intake m_intake = new Intake();
+  
+  private final IntakeRollers m_intakeRollers = new IntakeRollers();
+  private final IntakePivot m_intakePivot = new IntakePivot();
+  private final Intake m_intake = new Intake(m_intakeRollers, m_intakePivot);
 
   private final XboxController m_controller = new XboxController(0);
